Move LoadJsonDowngrades function to downgrader.cpp
Consolidates multiple copies of the LoadJsonDowngrades function into a
single version in downgrader.cpp. The Downgrader.h file, where it should
have been declared, included Buzz_inc.h because of pragma packing
issues. To avoid complications from that, two identical copies of the
LoadJsonDowngrade function were declared locally in rush.cpp and
newmis.cpp. Now, with global structure packing disabled by commit
f852876c7eb0fbcfa54d53ceb4d34e5beca8fccb, the declarations have been
consolidated in downgrader.h/.cpp (as is proper).
