+++ src/game/downgrader.cpp
@@ -1,9 +1,25 @@
-/* TODO: Copyright
+/* Copyright (C) 2020 Ryan Yoakum
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-#include <cassert>
+#include "downgrader.h"
 
-#include "downgrader.h" // Moved down because it includes Buzz_inc.h
+#include <cassert>
+#include <fstream>
+#include <json/json.h>
 
 #include "Buzz_inc.h"
 #include "ioexception.h"
@@ -181,4 +197,71 @@ struct MissionType Downgrader::next()
     }
 
     return mCurrent;
+}
+
+//----------------------------------------------------------------------
+// End of Downgrader class methods
+//----------------------------------------------------------------------
+
+
+/* Read the mission downgrade options from a file.
+ *
+ * The Json format is:
+ * {
+ *   "missions": [
+ *     { "mission": <Code>, "downgrades": [<Codes>] },
+ *     ...
+ *   ]
+ * }
+ *
+ * \param filename  A Json-formatted data file.
+ * \return  A collection of MissionType.MissionCode-indexed downgrade
+ *          options.
+ * \throws IOException  If filename is not a readable Json file.
+ */
+Downgrader::Options LoadJsonDowngrades(std::string filename)
+{
+    char *path = locate_file(filename.c_str(), FT_DATA);
+
+    if (path == NULL) {
+        free(path);
+        throw IOException(std::string("Unable to open path to ") +
+                          filename);
+    }
+
+    std::ifstream input(path);
+    Json::Value doc;
+    Json::Reader reader;
+    bool success = reader.parse(input, doc);
+
+    if (! success) {
+        free(path);
+        throw IOException("Unable to parse JSON input stream");
+    }
+
+    assert(doc.isObject());
+    Json::Value &missionList = doc["missions"];
+    assert(missionList.isArray());
+
+    Downgrader::Options options;
+
+    for (int i = 0; i < missionList.size(); i++) {
+        Json::Value &missionEntry = missionList[i];
+        assert(missionEntry.isObject());
+
+        int missionCode = missionEntry.get("mission", -1).asInt();
+        assert(missionCode >= 0);
+        // assert(missionCode >= 0 && missionCode <= 61);
+
+        Json::Value &codeGroup = missionEntry["downgrades"];
+        assert(codeGroup.isArray());
+
+        for (int j = 0; j < codeGroup.size(); j++) {
+            options.add(missionCode, codeGroup[j].asInt());
+        }
+    }
+
+    input.close();
+    free(path);
+    return options;
 }
